Update README for submodule
Now that toolforge_i18n is a submodule, the initial setup and updates
need to look a little different. (In practice, I expect I’ll usually
omit the `git submodule update` when I know there weren’t any submodule
changes, but that’s besides the point.)

Also fix the path in the setup – I changed the origin URL from GitHub to
GitLab in 52570991cd (and the GitLab repo doesn’t have the tool-
prefix), but neglected to update the `cd` command at the same time.