Use improved toolforge_i18n language_code_to_babel
And change the rest of our custom implementation to index the dict by
the the original language code and only return the “remove everything
after -” part as the final fallback, which feels safer – I wouldn’t want
to map, say, both aa-latn and aa-cyrl to the same aa fallback.

I think the improved upstream translation also means that we now map
zh-hans and zh-hant to zh_Hans and zh_Hant, respectively, rather than
both going to zh (and probably likewise for a few other language codes),
which seems nice (though I haven’t tested the overall impact of this).