Support opting back out of Wikifunctions mode
Previously, it wasn’t easy for users to opt back out of the experimental
Wikifunctions support, because users generally don’t have permission to
delete their own user pages. So instead of just checking for the
existence of the page, we now also check that the content isn’t the
exact string /* DISABLED */.

(At some point this should probably just become a preference in the
tool, to be honest. I think the feature is working well enough at this
point.)