Upgrade to toolforge_i18n 0.0.5
toolforge_i18n now checks translations by default when they are loaded;
since we run the translation checks in CI, we can skip this and speed up
the tool’s start-up time a bit. (Versions 0.0.3 and 0.0.4 failed to
publish to PyPI, by the way – this is really just one patch version
jump.)

Also, remove beautifulsoup4 from our dev dependencies – it was only
needed for the translation checks, and for that purpose is now pulled in
by toolforge_i18n in the main dependencies. (Since we disable the checks
at load time, we won’t actually use beautifulsoup4, but I don’t think
it’s worth jumping through hoops to exclude it from the Toolforge venv.)