Change Babel lanugage code of kaa (Karakalpak)
In commit ce4e46e621, I had mapped this to kk (Kazakh) but wasn’t happy
with the solution; now, the same language code came up in Wikidata Image
Positions, and (forgetting I’d already encountered the language code
here) I investigated the situation again and independently concluded
that uz (Uzbek) was the best fallback in wd-image-positions commit
242b58ce4a [1]; after realizing I’d previously chosen differently here,
I still think Uzbek is better than Kazakh: it’s not as closely related
to Karakalpak, but it’s at least still in the Turkic language family,
uses the right script (Latin), and has a chance of being intelligible to
Karakalpak speakers because Karakalpak is spoken in (an autonomous
republic) of Uzbekistan. So change it here to bring both tools in line.

[1]: https://gitlab.wikimedia.org/toolforge-repos/wd-image-positions/-/commit/242b58ce4a