Upgrade dependencies
Werkzeug 3.1.0 deprecates ImmutableOrderedMultiDict (along with
OrderedMultiDict), so stop using these two classes. We were using them
in two different ways:

First, as introduced in efb4c9f87e, we used ImmutableOrderedMultiDict
for the request storage (i.e. flask.request.args and similar). I now
believe this to be based on a misunderstanding: we need the order of the
multiple &form_representation= parameters to be preserved, but a regular
MultiDict does guarantee that; what an ImmutableOrderedMultiDict gives
you on top of that is merely that it preserves the order of *different*
keys, even if they are intermixed, but we don’t need that. This would
also explain why, as noted in that commit, “I really [didn’t] know why
this worked” – now I know.

Second, as introduced in 55af8047cf, we supported storing an
ImmutableOrderedMultiDict in the session, which was needed for
1070a3a8f2, where we stash the flask.request.args in the session (and,
per the above change, the args are an ImmutableOrderedMultiDict).
Without OrderedFlask, I don’t think this is necessary anymore. (I seem
to recall that this was also necessary for the /template/ API, back when
the templates were stored as an ordered dict between commits 4f0d8cc6cb
and 1f2a6f2e17, but I can’t find evidence for this in the Git log, and a
closer look at 4f0d8cc6cb also reveals that it used
collections.OrderedDict, not ImmutableOrderedMultiDict, so I might just
be mixing things up there. In any case it would not be needed anymore
since 1f2a6f2e17.) In theory, we might want to keep support for
ImmutableOrderedMultiDict in the session for a bit (to be able to parse
existing sessions), but as this support is only used for a very
short-lived session component (stash the args in the session, redirect
to URL without args, immediately pop the args out of the session again)
I don’t think that’s needed – let’s just deploy this and see if any
errors crop up.

Also, with reference to the previous commit, max_form_memory_size now
has a default value in Werkzeug. 500 kB should be plenty for us, so I
don’t see a need to change it.

    pip-compile -U && pip-compile -U dev-requirements.in