Add build service files
With this, it should be possible to build an image for the tool and run
it using the buildservice type, rather than python3.11. If it works,
I’ll update the documentation; first I need to push this so I can build
the image :)

I omitted the --bind setting from the gunicorn command because, if the
$PORT environment variable is set (which it should be on Toolforge), it
already defaults to 0.0.0.0:$PORT, which is just what we want.